Important changes to stmpy.load() for 3ds files.
1. stmpy.load() for 3ds now looks for the grid Z (the value recorded
during at the start of a dIdV sweep).  If it’s not found it will resort
to the scan Z, which is recored as the tip move backwards between lines
of the DOS map.

Note: this makes it possible to spot tip changes from the Z scan (was
previously harder)

2. New functions: stmpy.tools.interp2d().  This extends
spicy.interpolate.interp2d() to allow a kind=‘nearest’ keyword argument
(for no interpolation).

3. Changed stmpy.tools.arc_linecut() and stmpy.tools.linecut() to use
stmpy.tools.interp2d() for interpolation and explicitly take a ‘kind’
keyword argument.